ChatBots_Groq/requirements.txt
langchain
langchain-community
langchain_chroma
langchain-text-splitters
langchain_huggingface
langchain-openai
langchain_groq
streamlit

python-dotenv
ipykernel
pypdf
bs4
arxiv
pymupdf
Wikipedia

lxml
chromadb
sentence_transformers
faiss-cpu

